-Split UInt256 file.
